% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIPWsubtype.R
\name{AIPWsubtype}
\alias{AIPWsubtype}
\title{Augmented inverse probability weighted Cox proportional hazard model for competing risks data}
\usage{
AIPWsubtype(formula, data, id, missing_model, missing_indep = FALSE,
  two_stage = FALSE, tstage_name = NULL, marker_name,
  marker_rr = NULL, first_cont_rr = TRUE, second_cont_bl = FALSE,
  second_cont_rr = FALSE, constvar = NULL, init, control, x = FALSE,
  y = TRUE, model = FALSE)
}
\arguments{
\item{formula}{a formula object with an obect of the type \code{Surv} on the left side and the terms on the right side. It should not include marker variables.}

\item{data}{a data.frame which has variables in formula and markers.}

\item{id}{a charhacter string specifying subject IDs.}

\item{missing_model}{a list of the missingness model for each marker. A right side of formula object including a ~ operator. If \code{two_stage} is \code{TRUE}, a model for the first-stage missingness model should be included at the last element of a list.}

\item{missing_indep}{a logical value: if \code{TRUE}, markers are assumed to be independent.}

\item{two_stage}{a logical value: if \code{TRUE}, a two-stage missingness model is used. Default is \code{FALSE}.}

\item{tstage_name}{a charhacter string specifying the first-stage missingness variable if \code{two_stage} is \code{TRUE}.}

\item{marker_name}{a vector of charhacter strings specifying markers defining cause of failures.}

\item{marker_rr}{a vector of logical value. Dafault is NULL, in which a model includes all markers named in \code{marker_name} in modeling heterogeneity effects. Otherwise, a vector of logical value can spcify whether each marker's heterogeneity effect will be examined or not. A length of this should be equal to that of \code{marker_name}.}

\item{first_cont_rr}{a logical value: if \code{TRUE}, the first order contrasts are included in modeling cause-specific relative risks based on log-linear representation. Otherwise the first contrasts are only included.}

\item{second_cont_bl}{a logical value: if \code{TRUE}, the second order contrasts are included in modeling cause-specific baseline functions based on log-linear representation. Otherwise the first contrasts are only included.}

\item{second_cont_rr}{a logical value: if \code{TRUE}, the second order contrasts are included in modeling cause-specific relative risks based on log-linear representation. Otherwise the first contrasts are only included.}

\item{constvar}{a vector of character strings specifying constrained varaibles of which the effects on the outcome are to be the same across subtypes of outcome. The variables which are not specified in \code{constvar} are considered as unconstrained variabales of which the associations with the outcome may be different across the outcome subtypes.}

\item{init}{a vector of initial values of the iteration. Default value is zero for all variables.}

\item{control}{an object of class \code{\link[survival]{coxph.control}} in \code{survival} packages.The default value of \code{iter.max} is 2000 and that of \code{eps} is 1e-12. See \code{\link[survival]{coxph.control}} for other values.}
}
\value{
An object of class \code{AIPWsubtype} representing the fit.
\item{coefficients}{the estimated regressoin coefficients.}
\item{naive.var}{the inverse of estimated Information matrix.}
\item{var}{the robust sandwich variance estimate.}
\item{linear.predictors}{a vector of linear predictors. This is not centered.}
\item{loglik}{a vector of length 2 containing the log-likelihood with the initial values and with the estimated coefficients.}
\item{score}{a value of the score test at the initial value of the coefficients.}
\item{rscore}{a value of the robust log-rank statistic.}
\item{wald.test}{a value of the Wald test statistics for whether the estimated coefficients are different from the initial value of the coefficients.}
\item{score.residual}{the score residuals for each subject.}
\item{iter}{a number of iterations used.}
\item{conv}{an integer code for the convergence. 0 indicates successful convergence, 1 indicates a failure to convergence, and 2 indicates it reaches the maximum number of iterations.}
\item{basehaz}{estimated baseline cause-specific hazard functions the reference disease subtype corresponding to marker variables equal to 1.}
\item{Ithealp}{a matrix of the partial derivative of the score functions with respect to the parameters from the missingness models.}
\item{Ithegam}{a matrix of the partial derivative of the score functions with respect to the parameters from the marker model.}
\item{model_missing}{a list of an object of class \code{glm} fitting the missingness models.}
\item{model_subtype}{an object of class \code{clogit} fitting the marker model.}
\item{n}{the number of observations.}
\item{nevent}{the number of events.}
\item{subtype}{a list of values related to subtypes including the number of subtypes, character strings of marker names, etc.}


The object will also contain the following: strata, formula, call, terms, y, xlevels, offset, optionally x, and model.
}
\description{
Fitting an augmented inverse probability weighted Cox proportional hazard model for competing risks data with partially missing markers, in which marker variables define the subyptes of outcome.
}
\details{
The Cox proportional hazard model is used to model cause-specific hazard functions. To examine the association between exposures and the specific subtypes of disease, the log-linear is used for reparameterization. Logistic regression models \code{\link[stats]{glm}} are used for the missiness models and a conditional logistic regression model \code{\link[survival]{clogit}} is used for the marker model.
The data duplication method is used so that the returned value \code{x} and \code{y} are duplicated the number of subtypes of outcome. Special terms including \code{+strata()} and \code{+offset()} can be used. \code{+cluster()} should be avoided since we automatically include it in the formula. Breslow method is used for handling tied event times. The first order contrasts are included as default in modeling cause-specific baseline functions based on log-linear representation.

For marker variables, 0 indicates censored events or missing values.
formula = formula, call = Call, terms = Terms, assign = assign, method = "AIPW"
}
\examples{
m1 <- AIPWsubtype(Surv(start, time, status)~ X + W,  data = data, id = "id", missing_model = list(~time + X, ~time + X),
 two_stage = FALSE, marker_name = c("y1", "y2"), second_cont_bl = FALSE, second_cont_rr = FALSE, constvar = "W")

# Two-stage missingness model
m2 <- AIPWsubtype(Surv(start, time, status)~ X + W,  data = data, id = "id", missing_model = list(~time + X, ~time + X, ~time + X + W),
 two_stage = TRUE, tstage_name = c("R0"), marker_name = c("y1", "y2"), second_cont_bl = FALSE, second_cont_rr = FALSE, constvar = "W")


}
