% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIF.R
\name{cif}
\alias{cif}
\title{predicting the cumulative incidence function}
\usage{
cif(fit, newdata, individual, id, na.action = na.pass)
}
\arguments{
\item{fit}{\code{subtype}, \code{IPWsubtype} or \code{AIPWsubtype} object.}

\item{newdata}{a data.frame which has variables in formula and markers.}
}
\description{
predicting the cumulative incidence function for IPWsubtype and AIPWsubtype models.
}
