% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlList.R
\name{controlList}
\alias{controlList}
\title{Control parameters}
\usage{
controlList(control)
}
\arguments{
\item{eps}{Tolerance for convergence in the parameters. Default is 1e-12.}

\item{maxiter}{The maxiumum number of interations. Default is 2000.}
}
\description{
Control parameters
}
