% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetero.test.R
\name{hetero.test}
\alias{hetero.test}
\title{The heterogeneity test for the subtype-specific effet
Testing whether the exposures are not associated with the specific subtypes of disease.}
\usage{
hetero.test(fit, fit2 = NULL, data = NULL)
}
\arguments{
\item{fit}{an object of class \code{subtype}, \code{IPWsubtype}, or \code{AIPWsubtype}. If a single object \code{fit} was specified, it tests whether all unconstrained variables, so called, exposures are not associated with disease subtypes.}

\item{fit2}{an object of the same class as \code{fit} to compare the heterogeneity effect.}

\item{data}{a data frame which was used in \code{fit}. It must be specified if a single object \code{fit} is specified.}
}
\value{
a data frame: names of variables which are compared, test statistics, the degree of freedom, and p-value.
}
\description{
The heterogeneity test for the subtype-specific effet
Testing whether the exposures are not associated with the specific subtypes of disease.
}
\details{
Log-likelihood tests are conducted to test whether the heterogeneity effects are significant. To test the heterogeneity effects, \code{marker_rr} can be used to drop out the marker-specific effects of exposures.
}
