% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CIF.R
\name{plot.cif}
\alias{plot.cif}
\title{Plot the cumulative incidence function}
\usage{
\method{plot}{cif}(x, xlab = "TIME",
  ylab = "CUMULATIVE INCIDENCE FUNCTION", main = NULL, xlim = NULL,
  ylim = NULL, type = "l", lty = 1, lwd = 1, lend = par("lend"),
  pch = NULL, col = NULL, legend = FALSE, ...)
}
\arguments{
\item{x}{an object class \code{cif}.}

\item{xlab}{a x-axis label. Default is \code{"TIME"}.}

\item{ylab}{a y-axis label. Default is \code{"CUMULATIVE INCIDENCE FUNCTION"}.}

\item{main}{a main label.}

\item{xlim}{boundary for x values.}

\item{ylim}{boundary for y values.}

\item{type}{a type of plots from \code{\link[graphics]{matplot}}.}

\item{lty, lwd, lend, pch, col}{see \code{\link[graphics]{matplot}}.}

\item{legend}{a logical value: if \code{TRUE}, legend is created at the top-left corner of the plot.}
}
\description{
Plotting the cumulative incidence function for an object class \code{cif}.
}
