% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.hetero.test.R
\name{print.hetero.test}
\alias{print.hetero.test}
\title{Print heterogeneity test for the subtype-specific effet}
\usage{
\method{print}{hetero.test}(object)
}
\description{
Print heterogeneity test for the subtype-specific effet
}
\keyword{internal}
