% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtype.R
\name{subtype}
\alias{subtype}
\title{Cox proportional hazard model for competing risks data}
\usage{
subtype(formula, data, id, marker_name, second_cont_bl = FALSE,
  second_cont_rr = FALSE, constvar = NULL, init, control, x = FALSE,
  y = TRUE, model = FALSE, ...)
}
\arguments{
\item{formula}{a formula object with an obect of the type \code{Surv} on the left side and the terms on the right side. It should not include marker variables.}

\item{data}{a data.frame which has variables in formula and markers.}

\item{id}{a charhacter string specifying subject IDs.}

\item{marker_name}{a vector of charhacter strings specifying markers defining cause of failures.}

\item{second_cont_bl}{a logical value: if \code{TRUE}, the second order contrasts are included in modeling cause-specific baseline functions based on log-linear representation. Otherwise the first contrasts are only included.}

\item{second_cont_rr}{a logical value: if \code{TRUE}, the second order contrasts are included in modeling cause-specific relative risks based on log-linear representation. Otherwise the first contrasts are only included.}

\item{constvar}{a vector of character strings specifying constrained varaibles of which the effects on the outcome are to be the same across subtypes of outcome. The variables which are not specified in \code{constvar} are considered as unconstrained variabales of which the associations with the outcome may be different across the outcome subtypes.}

\item{init}{a vector of initial values of the iteration. Default value is zero for all variables.}

\item{control}{an object of class \code{\link[survival]{coxph.control}} in \code{survival} packages.}
}
\value{
A returned object is an object of class \code{\link[survival]{coxph}}. See \code{\link[survival]{coxph.object}} for details. The additional returned values include the following:
\item{basehaz}{estimated baseline cause-specific hazard functions the reference disease subtype corresponding to marker variables equal to 1.}
\item{subtype}{a list of values related to subtypes including the number of subtypes, character strings of marker names, etc.}
}
\description{
Fitting a Cox proportional hazard model for competing risks data, in which marker variables define the subyptes of outcome.
}
\details{
The Cox proportional hazard model is used to model cause-specific hazard functions. To examine the association between exposures and the specific subtypes of disease, the log-linear is used for reparameterization.
This is a wrapper function for \code{\link[survival]{coxph}} after data duplication. The returned value \code{x}, \code{y} are duplicated the number of subtypes of outcome. \code{+cluster()} should be avoided since we automatically include it in the formula.

For marker variables, 0 indicates censored events.
}
\examples{
m1 <- subtype(Surv(start, time, status)~ X + W,  data = data, id = "id", marker_name = c("y1", "y2"), second_cont_bl = FALSE, second_cont_rr = FALSE, constvar = "W")

}
